// components/types/Type.tsx
import { StaticImageData } from "next/image";

export type courseType = {
  heading: string;
  para: string;
};
export type tourType = {
  city: string;
  img: string;
};
export type faculityType = {
  img: string;
  heading: string;
  para: string;
};
export type blogType = {
  name: string;
  number: Number;
};